extracted enum ShakedownTestFilterTypes
